US923: Refactor of extension methods

+++ src/SFA.Apprenticeships.Infrastructure.Raa/Mappers/TraineeshipVacancyDetailMapper.cs
@@ -23,7 +23,7 @@
 
             var detail = new TraineeshipVacancyDetail
             {
-                Id = (int)vacancy.VacancyReferenceNumber,
+                Id = vacancy.VacancyReferenceNumber,
                 VacancyReference = vacancy.VacancyReferenceNumber.GetVacancyReference(),
                 Title = vacancy.Title,
                 Description = vacancy.ShortDescription,
